refactor: 공통 로직 Participant 분리
