+++ src/main/java/domain/Dealer.java
@@ -2,40 +2,12 @@ package domain;
 
 import java.util.List;
 
-public class Dealer {
-    private final List<Card> holdCards;
-
+public class Dealer extends Participant {
     public Dealer(List<Card> holdCards) {
-        this.holdCards = holdCards;
-    }
-
-    public void addCard(Card card) {
-        holdCards.add(card);
+        super(holdCards);
     }
 
     public boolean isReceiveCard() {
         return calculateTotalScore() <= 16;
-    }
-
-    public int calculateTotalScore() {
-        int results = 0;
-        for (Card holdCard : holdCards) {
-            results += holdCard.getScore();
-        }
-
-        boolean isAceExist = holdCards.stream()
-                .anyMatch(holdCard -> holdCard.getCardNumber() == CardNumber.ACE);
-        if (isAceExist && (results + 10) <= 21) {
-            return results + 10;
-        }
-        return results;
-    }
-
-    public List<Card> getHoldCards(){
-        return List.copyOf(holdCards);
-    }
-
-    public boolean isBust() {
-        return calculateTotalScore() > 21;
     }
 }
